backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
numpy==1.24.3
Pillow==10.1.0
cfgrib==0.9.10.4
xarray==2023.10.1
python-multipart==0.0.6

